requirements.txt
fredapi==0.5.2
pandas==2.2.3
matplotlib==3.9.3
seaborn==0.13.2
requests==2.32.3
python-telegram-bot==21.0
python-dotenv==1.0.0
